+++ js/eve-online-intel-tool.js
@@ -120,10 +120,13 @@ jQuery(document).ready(function($) {
 
 	/**
 	 * Sticky highlight similar table rows on click
+	 *
+	 * todo:
+	 * Get this to work with data tables pagination, so it's deactivated for now
 	 */
-	$('tr[data-highlight]').click(function() {
-		$('tr[data-highlight="' + $(this).data('highlight') + '"]').toggleClass('dataHighlightSticky');
-	});
+//	$('tr[data-highlight]').click(function() {
+//		$('tr[data-highlight="' + $(this).data('highlight') + '"]').toggleClass('dataHighlightSticky');
+//	});
 
 	/**
 	 * Getting the nonce for the form